Proof of concept
Using http://svgjs.com to draw a panel as a grid of modules.

The individual modules are defined as SVN symbols, and drawn using
SVG.use().  This is likely not good enough to change the indicator
colors via CSS, since each indicator will need it's own ID, or some
other attribute that is easily addressable from CSS.